Update DB configuration and DB schema
Update config to match mine, update schema to match the entity

+++ src/main/resources/schema.sql
@@ -1,8 +1,8 @@
 create table if not exists  product (
-    id INTEGER NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    id LONG NOT NULL AUTO_INCREMENT PRIMARY KEY,
     name VARCHAR(255),
     category VARCHAR(255),
     price DECIMAL(10,2),
-    quantity INTEGER
+    quantity LONG
     -- on update restrict
 );
